requirements.txt
fastapi 
uvicorn[standard] 
pydantic
pypdf
python-multipart
faiss-cpu 
sentence-transformers 
transformers